extract MiniAppParameters and MiniAppProgressView to UI module
